squash: Drops dispatching redirected action.
